Re-factoring of the single module project into a multi-module one
