+++ .travis.yml
@@ -1,0 +1,18 @@
+language: java
+jdk: oraclejdk8
+dist: trusty
+
+script: mvn -B verify -Dvaadin.productionMode -Dcom.vaadin.testbench.Parameters.testsInParallel=2 -Dvaadin.proKey=$VAADIN_PRO_KEY -Dsauce.user=$SAUCE_USER -Dsauce.sauceAccessKey=$SAUCE_ACCESSKEY
+
+# Cache requires a master build
+branches:
+  only:
+    - master
+
+cache:
+  directories:
+  - $HOME/.m2
+
+before_cache:
+  # remove all build artifacts
+  - rm -rf $HOME/.m2/repository/com/vaadin/vaadin-button-flow
