+++ vaadin-button-flow-testbench/src/license/cval3/header.txt
@@ -1,0 +1,8 @@
+This program is available under Commercial Vaadin Add-On License 3.0
+(CVALv3).
+
+See the file license.html distributed with this software for more
+information about licensing.
+
+You should have received a copy of the CVALv3 along with this program.
+If not, see <http://vaadin.com/license/cval-3>.
