+++ vaadin-button-flow-testbench/src/license/licenses.properties
@@ -1,0 +1,1 @@
+cval3=Commercial Vaadin Licence v3
