+++ vaadin-button-flow-testbench/src/main/java/com/vaadin/flow/component/button/testbench/ButtonElement.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2000-2018 Vaadin Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package com.vaadin.flow.component.button.testbench;
+
+import com.vaadin.testbench.TestBenchElement;
+import com.vaadin.testbench.elementsbase.Element;
+
+/**
+ * A TestBench element representing a <code>&lt;vaadin-button&gt;</code>
+ * element.
+ */
+@Element("vaadin-button")
+public class ButtonElement extends TestBenchElement {
+    @Override
+    public String getText() {
+        // The default implementation seems to use innerText, which adds a lot
+        // of whitespace in Edge
+        return getPropertyString("textContent");
+    }
+}
